Add client-playing custom sound, scriptable trigger
- Adds a ShipSound ScreenComponent to play sounds on clients
- Adds PlayerSpaceship:triggerCustomSound() and resetCustomSound()
  scripting functions to initiate and reset a ShipSound that plays
  on all crew stations.

This replicates two new pieces of data over the network to player
ships: a resources/-based path to an audio file, and a Boolean
recording whether the sound has been triggered. The audio file
must exist on the clients for clients to hear it.

Passing the relative path to an audio file to a PlayerSpaceship,
ie. `player:triggerCustomSound("sfx/laser.wav")`, signals that
the sound has been triggered. Clients check each tick whether its
ship's sound is triggered.

If so, the clients play the sound and flag it as having been
played on that crew screen. Playback is not guaranteed to be
simultaneous, and clients that join or (re-)connect while the
trigger is enabled, or who switch crew screens, will hear the
sound when the crew screen is loaded.

This design is meant to allow remote clients to hear audio that's
not part of the game, such as the scenario voice tracks in
resources/audio/scenario/... . The experience in a local game will
not be good, but those games also have `playSound()` available.

Once flagged, a given trigger won't play the sound again on that
crew position screen until sound is reset on the player ship.
Resetting the trigger on the player ship should also stop any
custom sound currently being played back.

This commit does _not_ implement this sound on main screens, nor
does it implement volume, pitch, or looping controls.

This feature should play the sound once per player, not screen,
and possibly should not play on client re-connection. Since it
mostly copies the functionality of ImpulseSound, the ImpulseSound
sould be reimplemented as a ShipSound.

Tested and confirmed to play on network-connected remote games.