[GuiAdvancedScrollText] Update max_prefix_width on entry change
onDraw renders the prefix of every entry in order to determine the
max prefix width. Most of these calculations are unnecessary
because the prefix width can change only when the entry list does,
and only when a new entry exceeds or removed entry had set the
max_prefix_width.

Persist max_prefix_width in the class and update it only when
adding an entry, or when removing an entry that has a prefix width
equal to the max_prefix_width.