Mark progress
With this branch's changes, performance improves when viewing the
log across the board, and especially improves when viewing the
bottom of the log. However, there are still bugs in rendering the
latest line of the log in the ShipsLogControl minimized state upon
loading the control (toggling fixes it), and the log now flickers
when resizing the window.

- Font preparing is moved out of onDraw to addEntry. To facilitate
  window resizing, the scroll list is cleared and re-added to
  re-trigger font prep, and positioning and the prepared font data
  are persisted and stored in the entry rather than calculated
  ephemerally in onDraw. This reduces font prep calls to only
  window size changes or log entry changes, rather than every frame.
- Lines that are below the rendered area are now ignored.