Add button to GM chat dialog to revert to scripted chat
Add a Script button to GM chat dialogs for entities with comms
scripts or callbacks. When clicked, the button closes the GM chat
and opens a standard chat channel from the player ship to the
entity with no hailing delay. This either launches any configured
scripted chat on the receiving entity or closes the channel.