Allow overriding of MouseRenderer visibility
MouseRenderer enforces cursor visibility while the cursor is in
the window. Allow that to be overridden so the cursor can be
intentionally hidden, for instance if SDL relative mouse mode is
active.

Also adds an isVisible() getter, replacing public access to the
access to the visible variable. MouseRenderer still enforces
cursor invisibility outside of the window.

This change moves responsibility for cursor visibility to screens
that attempt to change it.